split icon stories in preparation for the new style icons
