--- packages/common/src/components/icons/Cog.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import IconBase from 'react-icons/IconBase';
-
-export const CogIcon = props => (
-  <IconBase fill="none" viewBox="0 0 24 24" {...props}>
-    <path
-      clipRule="evenodd"
-      d="m13.0028 4h-2.4615l-.3702 1.48071c-.60683.16989-1.17922.42223-1.70433.74419l-1.25338-.75203-1.74056 1.74057.75204 1.2534c-.36472.59489-.6401 1.25046-.80748 1.94806l-1.41739.3543v2.4616l1.41741.3543c.16737.6976.44276 1.3532.80748 1.9481l-.75204 1.2534 1.74056 1.7406 1.2534-.7521c.59489.3647 1.25047.6401 1.94809.8075l.3543 1.4174h2.4616l.3543-1.4174c.6976-.1674 1.3532-.4428 1.9481-.8075l1.2534.752 1.7405-1.7405-.752-1.2534c.3647-.5949.6401-1.2505.8075-1.9481l1.4174-.3543v-2.4616l-1.4174-.3543c-.1674-.69762-.4427-1.35321-.8075-1.9481l.7521-1.25338-1.7406-1.74057-1.2534.75202c-.663-.40649-1.4014-.70202-2.1893-.86072zm-1.0028 11.4286c1.8936 0 3.4286-1.5351 3.4286-3.4286 0-1.8936-1.535-3.42858-3.4286-3.42858-1.8935 0-3.42856 1.53498-3.42856 3.42858 0 1.8935 1.53506 3.4286 3.42856 3.4286z"
-      fill="currentColor"
-      fillRule="evenodd"
-    />
-  </IconBase>
-);
