--- packages/common/src/components/icons/Curator.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import IconBase from 'react-icons/IconBase';
-
-export const CuratorIcon = props => (
-  <IconBase fill="none" viewBox="0 0 24 24" {...props}>
-    <path
-      d="m11.2374 7.49606c.0925 0 .1495-.05696.1637-.14951.2137-1.17486.2065-1.18909 1.4383-1.43119.0855-.01424.1496-.0712.1496-.17089 0-.09255-.0641-.14952-.1567-.16376-1.2247-.23496-1.1962-.25633-1.4312-1.43118-.0142-.09256-.0712-.14953-.1637-.14953-.0926 0-.1495.05697-.1638.14953-.235 1.16061-.1994 1.18197-1.43829 1.43118-.09256.01424-.14952.07121-.14952.16376 0 .09969.05696.15665.15665.17089 1.23176.2421 1.21756.25633 1.43116 1.43119.0143.09255.0712.14951.1638.14951zm-3.45337 4.98424c.14241 0 .25633-.0997.27058-.2421.24921-1.8869.30617-1.894 2.27139-2.26427.1353-.02135.2421-.12104.2421-.27056 0-.14241-.1068-.24922-.2421-.27058-1.95811-.27057-2.02218-.32753-2.27139-2.25001-.01425-.14953-.12105-.24922-.27058-.24922s-.2492.09969-.27056.25633c-.22785 1.87977-.32042 1.87977-2.27138 2.2429-.13528.02848-.24209.12817-.24209.27058 0 .15664.10681.24921.27769.27056 1.92248.30617 2.00793.36317 2.23578 2.24997.02136.1567.12103.2564.27056.2564zm4.93437 7.8679c.2065 0 .3631-.1495.3916-.3631.5198-3.8592 1.0609-4.4645 4.8774-4.8846.2279-.0284.3773-.1922.3773-.3987s-.1494-.3703-.3773-.3987c-3.7951-.4415-4.3362-1.0111-4.8774-4.88456-.0285-.2136-.1851-.36313-.3916-.36313s-.356.14953-.3916.36313c-.5127 3.85926-1.0609 4.46446-4.87742 4.88456-.22073.0284-.37738.1922-.37738.3987s.15665.3703.37738.3987c3.78802.5056 4.31492 1.0111 4.87742 4.8846.0356.2136.1922.3631.3916.3631z"
-      fill="currentColor"
-    />
-  </IconBase>
-);
