--- packages/common/src/components/icons/Dashboard.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import IconBase from 'react-icons/IconBase';
-
-export const DashboardIcon = props => (
-  <IconBase fill="none" viewBox="0 0 24 24" {...props}>
-    <path
-      clipRule="evenodd"
-      d="m4 3c-.55228 0-1 .44771-1 1v7.8999c0 .5523.44771 1 1 1h6.0999c.5523 0 1-.4477 1-1v-7.8999c0-.55228-.4477-1-1-1zm9.9 0c-.5523 0-1 .44771-1 1v4.29994c0 .55228.4477 1 1 1h6.0999c.5523 0 1-.44772 1-1v-4.29994c0-.55228-.4477-1-1-1zm-1 9.0999c0-.5522.4477-1 1-1h6.0999c.5523 0 1 .4478 1 1v7.8999c0 .5523-.4477 1-1 1h-6.0999c-.5523 0-1-.4477-1-1zm-8.9 2.5999c-.55228 0-1 .4477-1 1v4.2999c0 .5523.44771 1 1 1h6.0999c.5523 0 1-.4477 1-1v-4.2999c0-.5523-.4477-1-1-1z"
-      fill="currentColor"
-      fillRule="evenodd"
-    />
-  </IconBase>
-);
