--- packages/common/src/components/icons/Documentation.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import IconBase from 'react-icons/IconBase';
-
-export const DocumentationIcon = props => (
-  <IconBase fill="none" viewBox="0 0 24 24">
-    <g fill="currentColor">
-      <path d="m11.4707 19v-12.58019c0-3.35473-8.47058-2.09672-8.47058-1.57253l-.00001 12.05612c6.47751 0 5.35721 1.2778 8.47059 2.0966z" />
-      <path
-        d="m12.5293 19v-12.58019c0-3.35473 8.4706-2.09672 8.4706-1.57253v12.05612c-6.4775 0-5.3572 1.2778-8.4706 2.0966z"
-        fillOpacity=".4"
-      />
-    </g>
-  </IconBase>
-);
