--- packages/common/src/components/icons/Exit.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import IconBase from 'react-icons/IconBase';
-
-export const ExitIcon = props => (
-  <IconBase fill="none" viewBox="0 0 24 24" {...props}>
-    <path
-      clipRule="evenodd"
-      d="m17.1074 7h-2.1072-2.1072-1.6694c.3189-.59519.9469-1 1.6694-1h4.2144c1.0454 0 1.8928.84743 1.8928 1.89278v8.21442c0 1.0454-.8474 1.8928-1.8928 1.8928h-4.2144c-.7225 0-1.3505-.4048-1.6694-1h1.6694 2.1072 2.1072c.4931 0 .8928-.3997.8928-.8928v-8.21442c0-.49307-.3997-.89278-.8928-.89278zm-13.91397 5.4801 4.20282 4.321c.2579.2652.67605.2652.93395 0 .25791-.2651.25791-.695 0-.9602l-3.22644-3.3608 9.89624.0755v-1.1112h-9.89624l3.22644-3.2853c.25791-.26516.25791-.69507 0-.96023-.2579-.26516-.67605-.26516-.93395 0l-4.20282 4.32103c-.25791.2651-.25791.6951 0 .9602z"
-      fill="currentColor"
-      fillRule="evenodd"
-    />
-  </IconBase>
-);
