--- packages/common/src/components/icons/Patron.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import IconBase from 'react-icons/IconBase';
-
-export const PatronIcon = props => (
-  <IconBase viewBox="0 0 24 24" fill="none" {...props}>
-    <path d="M13 10.3L16 9L19 10.3V13.7L16 15L13 13.7V10.3Z" fill="#999999" />
-    <path d="M5 10.3L8 9L11 10.3V13.7L8 15L5 13.7V10.3Z" fill="#999999" />
-    <path
-      d="M12 4.65522L16.4615 7.74856V16.2514L12 19.3448L7.53846 16.2514V7.74856L12 4.65522Z"
-      fill="white"
-      stroke="white"
-      strokeWidth="1.07692"
-    />
-  </IconBase>
-);
