--- packages/common/src/components/icons/Pro.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import IconBase from 'react-icons/IconBase';
-
-export const ProIcon = props => (
-  <IconBase fill="none" viewBox="0 0 24 24" {...props}>
-    <rect fill="#535bcf" height="10" rx="1" width="24.2857" y="7" />
-    <path
-      d="m5.1685 15h1.38388v-2.1222h1.16335c1.50533 0 2.38107-.898 2.38107-2.2052 0-1.3008-.85976-2.21805-2.34591-2.21805h-2.58239zm1.38388-3.2312v-2.18286h.93324c.799 0 1.18572.43466 1.18572 1.08666 0 .6488-.38672 1.0962-1.17933 1.0962zm4.41772 3.2312h1.3615v-2.7773c0-.6041.441-1.0196 1.0419-1.0196.1886 0 .4474.032.5753.0735v-1.2081c-.1215-.0287-.2909-.0479-.4283-.0479-.5497 0-1.0003.3196-1.1793.9268h-.0512v-.8565h-1.3199zm5.6721.0959c1.4894 0 2.4162-1.0196 2.4162-2.5313 0-1.5213-.9268-2.5376-2.4162-2.5376-1.4893 0-2.4162 1.0163-2.4162 2.5376 0 1.5117.9269 2.5313 2.4162 2.5313zm.0064-1.0547c-.6871 0-1.0387-.6296-1.0387-1.4862 0-.8565.3516-1.4893 1.0387-1.4893.6744 0 1.0259.6328 1.0259 1.4893 0 .8566-.3515 1.4862-1.0259 1.4862z"
-      fill="#fff"
-    />
-  </IconBase>
-);
