--- packages/common/src/components/icons/Search.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import IconBase from 'react-icons/IconBase';
-
-export const SearchIcon = props => (
-  <IconBase fill="none" viewBox="0 0 24 24" {...props}>
-    <path
-      clipRule="evenodd"
-      d="m12 20c4.4183 0 8-3.5817 8-8 0-4.41828-3.5817-8-8-8-4.41828 0-8 3.58172-8 8 0 4.4183 3.58172 8 8 8zm-.3678-12c-2.00599 0-3.6322 1.62621-3.6322 3.6322 0 2.0061 1.62621 3.6323 3.6322 3.6323.7056 0 1.3649-.2016 1.9226-.55l2.2855 2.4452 1.1597-1.1597-2.2855-2.4452c.3484-.5577.55-1.217.55-1.9226 0-2.00599-1.6262-3.6322-3.6323-3.6322zm-1.99219 3.6322c0-1.1002.89199-1.99219 1.99219-1.99219 1.1003 0 1.9923.89199 1.9923 1.99219 0 1.1003-.892 1.9923-1.9923 1.9923-1.1002 0-1.99219-.892-1.99219-1.9923z"
-      fill="currentColor"
-      fillRule="evenodd"
-    />
-  </IconBase>
-);
