+++ packages/components/src/components/Icon/icons-new-style.tsx
@@ -1,0 +1,171 @@
+import React from 'react';
+
+import { Element } from '../Element';
+
+export const profile = props => (
+  <Element
+    as="svg"
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M9.42741 6.9928C9.6573 6.58174 9.85182 6.04216 9.85182 5.49568C9.85182 4.11736 8.98956 3 7.92591 3C7.58611 3 7.26686 3.11404 6.98971 3.31419C6.92064 3.29358 6.84832 3.27504 6.77261 3.25875C6.07718 3.16651 5.37906 4.16168 5.2133 5.48153C5.04755 6.80138 5.47693 7.94611 6.17236 8.03836C6.26082 7.99604 6.35037 7.95448 6.44018 7.91285C6.76753 8.24599 7.18023 8.44467 7.62875 8.44467C8.01829 8.44467 8.38082 8.29481 8.68387 8.03714C9.07841 8.05439 9.4859 8.03351 9.68107 7.83834C9.77042 7.74899 9.66682 7.57522 9.56342 7.40179C9.46923 7.24381 9.37522 7.08611 9.42741 6.9928Z"
+      fill="currentColor"
+    />
+    <path
+      d="M10.9799 13L4.02989 13C4.02989 13 3.44374 8.8998 7.46303 9.00187C11.4823 9.10395 10.9799 13 10.9799 13Z"
+      fill="currentColor"
+    />
+  </Element>
+);
+
+export const dashboard = props => (
+  <Element
+    as="svg"
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M3 3.40909C3 3.18316 3.18316 3 3.40909 3H7.09081C7.31675 3 7.4999 3.18316 7.4999 3.40909V12.5909C7.4999 12.8168 7.31675 13 7.09081 13H3.40909C3.18316 13 3 12.8168 3 12.5909V3.40909ZM8.5001 3.40909C8.5001 3.18316 8.68325 3 8.90919 3H12.5909C12.8168 3 13 3.18316 13 3.40909V6.09091C13 6.31684 12.8168 6.5 12.5909 6.5H8.90919C8.68325 6.5 8.5001 6.31684 8.5001 6.09091V3.40909ZM8.90919 7.5C8.68325 7.5 8.5001 7.68316 8.5001 7.90909V12.5909C8.5001 12.8168 8.68325 13 8.90919 13H12.5909C12.8168 13 13 12.8168 13 12.5909V7.90909C13 7.68316 12.8168 7.5 12.5909 7.5H8.90919Z"
+      fill="currentColor"
+    />
+  </Element>
+);
+
+export const documentation = props => (
+  <Element
+    as="svg"
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      d="M7.70605 12V5.21067C7.70605 3.40018 3.00018 4.0791 3.00018 4.362L3.00017 10.8685C6.59879 10.8685 5.9764 11.5581 7.70605 12Z"
+      fill="currentColor"
+    />
+    <path
+      d="M8.29395 12V5.21067C8.29395 3.40018 12.9998 4.0791 12.9998 4.362L12.9998 10.8685C9.40121 10.8685 10.0236 11.5581 8.29395 12Z"
+      fill="currentColor"
+    />
+  </Element>
+);
+
+export const searchBubble = props => (
+  <Element
+    as="svg"
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M8 13C10.7614 13 13 10.7614 13 8C13 5.23858 10.7614 3 8 3C5.23858 3 3 5.23858 3 8C3 10.7614 5.23858 13 8 13ZM7.77015 5.5C6.51638 5.5 5.5 6.51638 5.5 7.77015C5.5 9.02392 6.51638 10.0403 7.77015 10.0403C8.21112 10.0403 8.62321 9.91432 8.97175 9.69654L10.4002 11.2248L11.125 10.5L9.69654 8.97175C9.91432 8.62321 10.0403 8.21112 10.0403 7.77015C10.0403 6.51638 9.02392 5.5 7.77015 5.5ZM6.52501 7.77015C6.52501 7.08248 7.08248 6.52501 7.77015 6.52501C8.45783 6.52501 9.0153 7.08248 9.0153 7.77015C9.0153 8.45783 8.45783 9.0153 7.77015 9.0153C7.08248 9.0153 6.52501 8.45783 6.52501 7.77015Z"
+      fill="currentColor"
+    />
+  </Element>
+);
+
+export const patron = props => (
+  <Element
+    as="svg"
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M8.28451 4.43355L10.3502 5.32691L11.7068 4.70781L8.28451 3V4.43355ZM4 10.8712L5.36265 10.2492V5.75087L4 5.12896V10.8712ZM12 10.8712L10.6372 10.2492V5.75087L12 5.12896V10.8712ZM5.65105 5.32691L7.71689 4.43355V3L4.29464 4.70781L5.65105 5.32691ZM7.71669 13L4.29443 11.2922L5.65084 10.673L7.71669 11.5665V13ZM8.28467 11.5665V13L11.7069 11.2922L10.3504 10.673L8.28467 11.5665ZM7.99992 4.85962L10.069 5.75446V10.2454L7.99992 11.1404L5.93066 10.2454V5.75446L7.99992 4.85962Z"
+      fill="currentColor"
+    />
+  </Element>
+);
+
+export const proBadge = props => (
+  <Element
+    as="svg"
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M3.35714 5.85718C3.1599 5.85718 3 6.01708 3 6.21432V9.78575C3 9.98299 3.1599 10.1429 3.35714 10.1429H12.6429C12.8401 10.1429 13 9.98299 13 9.78575V6.21432C13 6.01708 12.8401 5.85718 12.6429 5.85718H3.35714ZM5.16802 9.57139H4.64083V7.07788H5.62459C6.19075 7.07788 6.51826 7.42732 6.51826 7.92285C6.51826 8.42082 6.18466 8.76295 5.6112 8.76295H5.16802V9.57139ZM5.16802 7.50889V8.34047H5.52597C5.82792 8.34047 5.97524 8.17001 5.97524 7.92285C5.97524 7.67448 5.82792 7.50889 5.52354 7.50889H5.16802ZM7.38789 9.57139H6.86069V7.07788H7.84446C8.41061 7.07788 8.73813 7.39688 8.73813 7.89241C8.73813 8.23332 8.58107 8.48048 8.29738 8.60223L8.82701 9.57139H8.24503L7.77263 8.68746H7.38789V9.57139ZM7.38789 7.50889V8.26376H7.74584C8.04779 8.26376 8.19511 8.13957 8.19511 7.89241C8.19511 7.64404 8.04779 7.50889 7.7434 7.50889H7.38789ZM10.2345 7.04379C10.8956 7.04379 11.4021 7.50889 11.4021 8.32464C11.4021 9.14038 10.8956 9.60548 10.2345 9.60548C9.5697 9.60548 9.06686 9.13673 9.06686 8.32464C9.06686 7.50889 9.5697 7.04379 10.2345 7.04379ZM10.2345 9.13917C10.6168 9.13917 10.8676 8.85305 10.8676 8.32464C10.8676 7.79623 10.6168 7.51011 10.2345 7.51011C9.85095 7.51011 9.60136 7.79623 9.60136 8.32464C9.60136 8.85305 9.85095 9.13917 10.2345 9.13917Z"
+      fill="currentColor"
+    />
+  </Element>
+);
+
+export const curator = props => (
+  <Element
+    as="svg"
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      d="M7.69687 4.99273C7.75914 4.99273 7.79748 4.95441 7.80705 4.89213C7.95079 4.10162 7.94599 4.09204 8.77482 3.92915C8.83231 3.91956 8.87543 3.88124 8.87543 3.81416C8.87543 3.75189 8.83231 3.71355 8.77004 3.70397C7.94599 3.54588 7.96516 3.5315 7.80705 2.74099C7.79748 2.67872 7.75914 2.64038 7.69687 2.64038C7.63458 2.64038 7.59625 2.67872 7.58667 2.74099C7.42857 3.52192 7.45252 3.53629 6.6189 3.70397C6.55662 3.71355 6.5183 3.75189 6.5183 3.81416C6.5183 3.88124 6.55662 3.91956 6.62369 3.92915C7.45252 4.09204 7.44294 4.10162 7.58667 4.89213C7.59625 4.95441 7.63458 4.99273 7.69687 4.99273ZM5.37325 8.34639C5.46908 8.34639 5.54573 8.27932 5.55531 8.1835C5.723 6.9139 5.76132 6.90911 7.08362 6.65998C7.17465 6.64561 7.24651 6.57854 7.24651 6.47793C7.24651 6.38211 7.17465 6.31025 7.08362 6.29587C5.76611 6.11381 5.723 6.07549 5.55531 4.78193C5.54573 4.68132 5.47386 4.61425 5.37325 4.61425C5.27264 4.61425 5.20558 4.68132 5.1912 4.78672C5.03789 6.05154 4.97561 6.05154 3.66289 6.29587C3.57187 6.31503 3.5 6.38211 3.5 6.47793C3.5 6.58333 3.57187 6.64561 3.68684 6.65998C4.9804 6.86599 5.03789 6.90432 5.1912 8.17392C5.20558 8.27932 5.27264 8.34639 5.37325 8.34639ZM8.69338 13.6404C8.83231 13.6404 8.93772 13.5398 8.95688 13.396C9.30662 10.7994 9.67073 10.3921 12.2387 10.1095C12.392 10.0903 12.4926 9.9801 12.4926 9.84117C12.4926 9.70223 12.392 9.59203 12.2387 9.57287C9.6851 9.27584 9.32099 8.89256 8.95688 6.28629C8.93772 6.14256 8.83231 6.04195 8.69338 6.04195C8.55444 6.04195 8.45383 6.14256 8.42988 6.28629C8.08493 8.88298 7.71603 9.29021 5.14808 9.57287C4.99956 9.59203 4.89416 9.70223 4.89416 9.84117C4.89416 9.9801 4.99956 10.0903 5.14808 10.1095C7.69687 10.4496 8.0514 10.7898 8.42988 13.396C8.45383 13.5398 8.55923 13.6404 8.69338 13.6404Z"
+      fill="currentColor"
+    />
+  </Element>
+);
+
+export const feedback = props => (
+  <Element
+    as="svg"
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M2.99979 7.93502C2.9996 7.92157 2.99951 7.90809 2.99951 7.89459C2.99951 6.29584 4.29556 4.99979 5.89432 4.99979C7.49308 4.99979 8.78913 6.29584 8.78913 7.89459C8.78913 9.49335 7.49308 10.7894 5.89432 10.7894C5.8897 10.7894 5.88508 10.7894 5.88047 10.7894H2.99979V7.93502Z"
+      fill="currentColor"
+    />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M8.08643 10.5218C8.83793 9.89426 9.31606 8.95029 9.31606 7.89465C9.31606 7.53839 9.2616 7.19485 9.16056 6.8719C9.52234 6.68463 9.93312 6.57886 10.3686 6.57886C11.822 6.57886 13.0002 7.75708 13.0002 9.2105V9.21057H13.0004V11.8422H10.3688V11.8421H10.3686C9.39266 11.8421 8.54081 11.3109 8.08643 10.5218Z"
+      fill="currentColor"
+    />
+  </Element>
+);
+
+export const folder = props => (
+  <Element
+    as="svg"
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      d="M6.63641 4H3.45455C3.20351 4 3 4.18851 3 4.42105V11.5789C3 11.8115 3.20351 12 3.45455 12H12.5455C12.7965 12 13 11.8115 13 11.5789V5.40351C13 5.17097 12.7965 4.98246 12.5455 4.98246H8L6.93878 4.10667C6.85553 4.03796 6.74791 4 6.63641 4Z"
+      fill="currentColor"
+    />
+  </Element>
+);
+
+export const signout = props => (
+  <Element
+    as="svg"
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M11.0536 5.64037H10H8.94644H8.11175C8.2712 5.34278 8.58519 5.14038 8.94644 5.14038H11.0536C11.5763 5.14038 12 5.56409 12 6.08676V10.1939C12 10.7166 11.5763 11.1403 11.0536 11.1403H8.94644C8.58519 11.1403 8.2712 10.9379 8.11175 10.6403H8.94644H10H11.0536C11.3002 10.6403 11.5 10.4405 11.5 10.1939V6.08676C11.5 5.84023 11.3002 5.64037 11.0536 5.64037ZM4.09672 8.38039L6.1981 10.5409C6.32705 10.6735 6.53612 10.6735 6.66507 10.5409C6.79402 10.4083 6.79402 10.1934 6.66507 10.0608L5.05187 8.38039L9.99993 8.41814V7.86255H5.05187L6.66507 6.21992C6.79402 6.08734 6.79402 5.87239 6.66507 5.73981C6.53612 5.60723 6.32705 5.60723 6.1981 5.73981L4.09672 7.9003C3.96776 8.03284 3.96776 8.24784 4.09672 8.38039Z"
+      fill="currentColor"
+    />
+  </Element>
+);
