+++ packages/components/src/components/Icon/index.stories.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
-import { Icon } from '.';
+import { Icon, IconNames } from '.';
 import { SidebarRow, Text, Stack } from '../..';
 import * as icons from './icons';
+import * as iconsNewStyle from './icons-new-style';
 
 export default {
   title: 'components/Icon',
@@ -23,9 +24,23 @@ export const Comments = () => (
   </SidebarRow>
 );
 
+export const OldVsNewStyle = () => (
+  <Stack direction="vertical" gap={2}>
+    <Text>
+      The old style (left) icons are fitted into a 16x16 grid. The new style
+      icons (right) are exported with their proportional bounding boxes at
+      16x16, so they align visually with the font size.
+    </Text>
+    <Stack gap={2}>
+      <Icon name="github" size={32} css={{ border: '1px solid red' }} />
+      <Icon name="folder" size={32} css={{ border: '1px solid red' }} />
+    </Stack>
+  </Stack>
+);
+
 export const List = () => (
   <Stack css={{ flexWrap: 'wrap' }}>
-    {Object.keys(icons).map(name => (
+    {Object.keys(icons).map((name: IconNames) => (
       <Stack
         key={name}
         direction="vertical"
@@ -33,9 +48,26 @@ export const List = () => (
         gap={2}
         css={{ width: 64, margin: 4 }}
       >
-        {/* 
-        // @ts-ignore you would only do this in a story, never in the app */}
         <Icon name={name} />
+        <Text size={2} variant="muted">
+          {name}
+        </Text>
+      </Stack>
+    ))}
+  </Stack>
+);
+
+export const ListNewStyle = () => (
+  <Stack css={{ flexWrap: 'wrap' }}>
+    {Object.keys(iconsNewStyle).map((name: IconNames) => (
+      <Stack
+        key={name}
+        direction="vertical"
+        align="center"
+        gap={1}
+        css={{ width: 64, margin: 4, marginRight: 0 }}
+      >
+        <Icon name={name} size={24} />
         <Text size={2} variant="muted">
           {name}
         </Text>
